+++ src/app/app.component.html
@@ -3,7 +3,7 @@
 <div id="mainContainer" class="container-fluid col-sm-12 h-100">
   
   <div class="row">
-    <div class="col-sm-9 h-100">
+    <div class="col-sm-9 h-100 p-0">
         <app-main-column></app-main-column>
     </div>
     <div class="col-sm-3 h-100 d-flex justify-content-center">
