+++ src/app/comuns/class/classesComuns.ts
