+++ src/app/mainmenu/mainmenu.component.html
@@ -1,4 +1,4 @@
-<nav class="navbar navbar-expand-lg navbar-dark col-sm-12 mb-0">
+<nav id="top_main_menu" class="navbar navbar-expand-lg col-sm-12 mb-0">
   <a class="navbar-brand" [routerLink]="['/projeto']">
     <h3 class="mx-1 my-1">Vinícius Silvano</h3>
   </a>
