+++ src/app/projetos-carousel-imagens/projetos-carousel-imagens.component.html
@@ -1,17 +1,19 @@
 <!--<p *ngIf="alert">
   <ngb-alert [type]="alert.type" (close)="close(alert)">{{ alert.message }}</ngb-alert>
 </p>-->
+
 <ngb-carousel class="carouselContainer my-4" *ngIf="projeto.imagens">
-    <ng-template ngbSlide *ngFor="let imagem of projeto.imagens">
-      <div  (click)="ampliarImagemAtual(imagem,content)">
-        <img class="carouselImage text-white" [src]="imagem.source" alt="{{imagem.alt}}">
-        <div class="carousel-caption ">
-        <h3 *ngIf="imagem.title" id="title" [innerHTML]="imagem.title"></h3>
-        <p *ngIf="imagem.descricaoCarousel" id="descricaoCarousel" [innerHTML]="imagem.descricaoCarousel"></p>
-        </div>
+  <ng-template ngbSlide *ngFor="let imagem of projeto.imagens">
+    <div  (click)="ampliarImagemAtual(imagem,content)">
+      <img class="carouselImage text-white" [src]="imagem.source" alt="{{imagem.alt}}">
+      <div class="carousel-caption ">
+      <h3 [hidden]="esconderTitle" *ngIf="imagem.title" id="title" [innerHTML]="imagem.title"></h3>
+      <p [hidden]="esconderDescricao" *ngIf="imagem.descricaoCarousel" id="descricaoCarousel" [innerHTML]="imagem.descricaoCarousel"></p>
       </div>
-    </ng-template>
+    </div>
+  </ng-template>
 </ngb-carousel>
+
 
 
   <!--<ng-template #content id="imagensProjetoModal" let-modal>
