+++ src/app/projetos-especificacao/projetos-especificacao.component.css
@@ -5,3 +5,4 @@
 .accordionCard{
     color: black;
 }
+
