+++ src/app/right-aside-column/right-aside-column.component.html
@@ -1,12 +1,12 @@
-<aside class="">
-  <div  class="row ml-2">
-    <div id="crachaDeApresentacao" class="card text-dark">
-      <div class="card-body">
+<aside class="row">
+  <div id="crachaDeApresentacao" class="card">
+    <div class="card-body">
+      <article>
         <h2 class="card-title">Apresentação</h2>
         <img height="50%" width="100%" src="../..{{globalUrl}}/assets/images/minhaFoto.jpg" alt="Foto de Vinícius Silvano"/>
         <p class="card-text apresentacaoCardText">Olá meu nome é Vinícius Silvano e bem vindo ao meu portifólio, neste site você poderá encontrar informações e, possivelmente, artefatos de alguns projetos que eu desenvolvi.</p>
         <p class="card-text apresentacaoCardText">Atenciosamente Vinícius</p>
-       </div>
+      </article>
     </div>
   </div>
 </aside>
