+++ src/styles.css
@@ -11,22 +11,31 @@
 :root{
     --cor-azul-marinho: rgb(40,62,74);
     --cor-cinza-escuro: #3f3f3f;
-    --cor-preto-escuro: #323232;
+    --cor-preto-escuro: #121212;
+    --cor-preto-escuro-n2: #323232;
+    --cor-preto-default: black;
     --cor-verde-claro: #2F5D61;
     --cor-azul-claro: #000FFF;
     --cor-cinza-claro: #CFD3D4;
+    --cor-verde-claro-n2: #15BE14;
 }
 
 body{
     background-repeat: no-repeat;
     background-position: center;
     background-size: cover;
-    background-color: var(--cor-cinza-escuro);
-    color: var(--cor-cinza-claro)
+    background-color: var(--cor-preto-escuro);
+    color: var(--cor-verde-claro-n2)
 }
 
 #mainColumnCard{
     background-color: var(--cor-preto-escuro);
+    border-right-color: var(--cor-cinza-escuro);
+    border-right-style: solid;
+    border-right-width: small;
+    border-bottom-color: var(--cor-cinza-escuro);
+    border-bottom-style: solid;
+    border-bottom-width: small;
 }
 
 
@@ -46,16 +55,6 @@ body{
   background-color: var(--cor-preto-escuro);
 }
 
-.innerCardContainer .card ngb-carousel{
-  width: 100%;
-  height: 500px;
-}
-
-.innerCardContainer .card ngb-carousel img{
-  width: 100%;
-  height: 500px;
-}
-
 .breadcrumbContainer{
   padding-left: 3%;
   margin-bottom: 3%;
@@ -69,5 +68,4 @@ body{
 .carousel-control-prev-icon, .carousel-control-next-icon{
   background-color: var(--cor-azul-claro);
 }
-
